requirements.txt
google-cloud-bigquery
PyYAML
numpy